sanket_binance_bot/requirements.txt
python-binance==1.0.29
python-dotenv==1.1.1
requests>=2.32.0
websockets>=15.0.0
aiohttp>=3.12.0
pycryptodome>=3.23.0
